+++ src/predicates/mod.rs
@@ -1,8 +1,10 @@
 mod expansion;
+mod incircle;
 mod orient2d;
 mod orient3d;
 mod sign;
 
+pub use incircle::incircle;
 pub use orient2d::orient2d;
 pub use orient3d::orient3d;
 pub use sign::{Orientation, Sign};